feat: create student service
