+++ src/app/components/students/student-page/student-page.component.html
@@ -4,5 +4,5 @@
     <fa-icon [icon]="plusIcon"></fa-icon>
   </app-button>
 </div>
-<app-student-form (submit)="addData($event)">
+<app-student-form (submit)="addStudent($event)">
 </app-student-form>
